requirements.txt
# NemoClaw Gateway Dashboard - Python Dependencies
# Phase 1: Foundation & Engineer View (MVP)

# Web Framework
streamlit>=1.30.0
streamlit-elements>=0.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.18.0
altair>=5.2.0

# System Integration
psutil>=5.9.0
pynvml>=11.5.0
watchdog>=3.0.0
paramiko>=3.4.0

# Configuration & Validation
pyyaml>=6.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0
humanize>=4.9.0
tabulate>=0.9.0
tenacity>=8.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
